+++ src/components/Filters.svelte
@@ -10,14 +10,14 @@
 
     let showFilters = false;
     let choices = possibilities.map((choice, i) => ({
-        choice, selected: i === 0
+        choice,
+        selected: i === 0,
     }));
 
-    const handleClick = index => () => {
-        const activeCount = choices.filter(choice => choice.selected).length;
+    const handleClick = (index) => () => {
+        const activeCount = choices.filter((choice) => choice.selected).length;
         const choice = choices[index];
-        if (activeCount <= 1 && choice.selected === true)
-            return
+        if (activeCount <= 1 && choice.selected === true) return;
 
         choice.selected = !choice.selected;
         choices = choices;
@@ -25,7 +25,7 @@
 
     $: {
         const selectedChoices = choices.filter(({ selected }) => selected);
-        const mappedChoices = selectedChoices.map(({ choice }) => choice)
+        const mappedChoices = selectedChoices.map(({ choice }) => choice);
         dispatch("change", mappedChoices);
     }
 
@@ -35,21 +35,21 @@
 </script>
 
 <div class="filters">
-    <Button on:click={() => showFilters = !showFilters}>
-        Show Filters
+    <Button on:click={() => (showFilters = !showFilters)}>
+        {showFilters ? 'Cacher' : 'Montrer'}
+        les Filtres
     </Button>
     {#if showFilters}
-    <ul transition:slide class="filters__items">
-        {#each choices as { choice, selected }, index}
-        <li
-            class="filters__item"
-            class:-selected={selected}
-            on:click={handleClick(index)}
-        >
-            {choice}
-        </li>
-        {/each}
-    </ul>
+        <ul transition:slide class="filters__items">
+            {#each choices as { choice, selected }, index}
+                <li
+                    class="filters__item"
+                    class:-selected={selected}
+                    on:click={handleClick(index)}>
+                    {choice}
+                </li>
+            {/each}
+        </ul>
     {/if}
 </div>
 
@@ -60,10 +60,12 @@
         top: 1em;
         left: 1em;
 
+        z-index: 25;
+
         &__items {
             padding: 1.5em;
-            border-radius: .5em;
-            box-shadow: 0 .5em 1.5em rgba(black, 0.15);
+            border-radius: 0.5em;
+            box-shadow: 0 0.5em 1.5em rgba(black, 0.15);
 
             background-color: white;
         }
@@ -72,28 +74,32 @@
             cursor: pointer;
 
             user-select: none;
-            background-color: rgba(orange, .25);
-            border: .15em solid transparent;
+            background-color: rgba(orange, 0.25);
+            border: 0.15em solid transparent;
 
-            padding: .25em 1em;
-            border-radius: .25em;
+            padding: 0.25em 1em;
+            border-radius: 0.25em;
             transform-origin: center left;
 
             transition-duration: 200ms;
             transition-property: background-color border-color transform color;
 
-            &:not(:last-child) { margin-bottom: .5em }
             &:hover {
                 background-color: transparent;
-                border-color:  #ff3e00;
+                border-color: #ff3e00;
                 transform: scale(1.05);
             }
-            &:active { transform: scale(1) }
+
+            &:not(:last-child) { margin-bottom: 0.5em; }
+            &:active { transform: scale(1); }
 
             &.-selected {
                 background-color: #ff3e00;
                 color: white;
             }
+
+            &:first-child { transform-origin: left top }
+            &:last-child { transform-origin: left bottom }
         }
     }
 </style>