+++ src/kana/hiragana.js
@@ -1,0 +1,58 @@
+export default [
+    { word: "あ",   answer: "a" },
+    { word: "い",   answer: "i" },
+    { word: "う",   answer: "u" },
+    { word: "え",   answer: "e" },
+    { word: "お",   answer: "o" },
+
+    { word: "か",   answer: "ka" },
+    { word: "き",   answer: "ki" },
+    { word: "く",   answer: "ku" },
+    { word: "け",   answer: "ke" },
+    { word: "こ",   answer: "ko" },
+
+    { word: "さ",   answer: "sa" },
+    { word: "し",   answer: "shi" },
+    { word: "す",   answer: "su" },
+    { word: "せ",   answer: "se" },
+    { word: "そ",   answer: "so" },
+
+    { word: "た",   answer: "ta" },
+    { word: "ち",   answer: "chi" },
+    { word: "つ",   answer: "tsu" },
+    { word: "て",   answer: "te" },
+    { word: "と",   answer: "to" },
+
+    { word: "な",   answer: "na" },
+    { word: "に",   answer: "ni" },
+    { word: "ぬ",   answer: "nu" },
+    { word: "ね",   answer: "ne" },
+    { word: "の",   answer: "no" },
+
+    { word: "は",   answer: "ha" },
+    { word: "ひ",   answer: "hi" },
+    { word: "ふ",   answer: "fu" },
+    { word: "へ",   answer: "he" },
+    { word: "ほ",   answer: "ho" },
+
+    { word: "ま",   answer: "ma" },
+    { word: "み",   answer: "mi" },
+    { word: "む",   answer: "mu" },
+    { word: "め",   answer: "me" },
+    { word: "も",   answer: "mo" },
+
+    { word: "や",   answer: "ya" },
+    { word: "ゆ",   answer: "yu" },
+    { word: "よ",   answer: "yo" },
+
+    { word: "ら",   answer: "ra" },
+    { word: "り",   answer: "ri" },
+    { word: "る",   answer: "ru" },
+    { word: "れ",   answer: "re" },
+    { word: "ろ",   answer: "ro" },
+
+    { word: "わ",   answer: "wa" },
+    { word: "を",   answer: "wo" },
+
+    { word: "ん",   answer: "n" },
+]