+++ src/kana/hiraganaAccents.js
@@ -1,0 +1,31 @@
+export default [
+    { word: "が",   answer: "ga" },
+    { word: "ぎ",   answer: "gi" },
+    { word: "ぐ",   answer: "gu" },
+    { word: "げ",   answer: "ge" },
+    { word: "ご",   answer: "go" },
+
+    { word: "ざ",   answer: "za" },
+    { word: "じ",   answer: "ji" },
+    { word: "ず",   answer: "zu" },
+    { word: "ぜ",   answer: "ze" },
+    { word: "ぞ",   answer: "zo" },
+
+    { word: "だ",   answer: "da" },
+    { word: "ぢ",   answer: "ji" },
+    { word: "づ",   answer: "zu" },
+    { word: "で",   answer: "de" },
+    { word: "ど",   answer: "do" },
+
+    { word: "ば",   answer: "ba" },
+    { word: "び",   answer: "bi" },
+    { word: "ぶ",   answer: "bu" },
+    { word: "べ",   answer: "be" },
+    { word: "ぼ",   answer: "bo" },
+
+    { word: "ぱ",   answer: "pa" },
+    { word: "ぴ",   answer: "pi" },
+    { word: "ぷ",   answer: "pu" },
+    { word: "ぺ",   answer: "pe" },
+    { word: "ぽ",   answer: "po" },
+]