+++ src/kana/index.js
@@ -1,0 +1,15 @@
+import hiragana from "./hiragana";
+import hiraganaAccents from "./hiraganaAccents";
+import katakana from "./katakana";
+import katakanaAccents from "./katakanaAccents";
+import { hiraganaCombinations, katakanaCombinations } from "./combinations";
+
+export default {
+    "Hiragana": hiragana,
+    "Hiragana Accents": hiraganaAccents,
+    "Les Combinaisons Hiragana": hiraganaCombinations,
+
+    "Katakana": katakana,
+    "Katakana Accents": katakanaAccents,
+    "Les Combinaisons Katakana": katakanaCombinations,
+};