+++ src/kana/katakana.js
@@ -1,0 +1,58 @@
+export default [
+    { word: "ア",   answer: "a" },
+    { word: "イ",   answer: "i" },
+    { word: "ウ",   answer: "u" },
+    { word: "エ",   answer: "e" },
+    { word: "オ",   answer: "o" },
+
+    { word: "カ",   answer: "ka" },
+    { word: "キ",   answer: "ki" },
+    { word: "ク",   answer: "ku" },
+    { word: "ケ",   answer: "ke" },
+    { word: "コ",   answer: "ko" },
+
+    { word: "サ",   answer: "sa" },
+    { word: "シ",   answer: "shi" },
+    { word: "ス",   answer: "su" },
+    { word: "セ",   answer: "se" },
+    { word: "ソ",   answer: "so" },
+
+    { word: "タ",   answer: "ta" },
+    { word: "チ",   answer: "chi" },
+    { word: "ツ",   answer: "tsu" },
+    { word: "テ",   answer: "te" },
+    { word: "ト",   answer: "to" },
+
+    { word: "ナ",   answer: "na" },
+    { word: "ニ",   answer: "ni" },
+    { word: "ヌ",   answer: "nu" },
+    { word: "ネ",   answer: "ne" },
+    { word: "ノ",   answer: "no" },
+
+    { word: "ハ",   answer: "ha" },
+    { word: "ヒ",   answer: "hi" },
+    { word: "フ",   answer: "fu" },
+    { word: "ヘ",   answer: "he" },
+    { word: "ホ",   answer: "ho" },
+
+    { word: "マ",   answer: "ma" },
+    { word: "ミ",   answer: "mi" },
+    { word: "ム",   answer: "mu" },
+    { word: "メ",   answer: "me" },
+    { word: "モ",   answer: "mo" },
+
+    { word: "ヤ",   answer: "ya" },
+    { word: "ユ",   answer: "yu" },
+    { word: "ヨ",   answer: "yo" },
+
+    { word: "ラ",   answer: "ra" },
+    { word: "リ",   answer: "ri" },
+    { word: "ル",   answer: "ru" },
+    { word: "レ",   answer: "re" },
+    { word: "ロ",   answer: "ro" },
+
+    { word: "ワ",   answer: "wa" },
+    { word: "ヲ",   answer: "wo" },
+
+    { word: "ン",   answer: "n" },
+]