+++ src/kana/katakanaAccents.js
@@ -1,0 +1,31 @@
+export default [
+    { word: "ガ",   answer: "ga" },
+    { word: "ギ",   answer: "gi" },
+    { word: "グ",   answer: "gu" },
+    { word: "ゲ",   answer: "ge" },
+    { word: "ゴ",   answer: "go" },
+
+    { word: "ザ",   answer: "za" },
+    { word: "ジ",   answer: "ji" },
+    { word: "ズ",   answer: "zu" },
+    { word: "ゼ",   answer: "ze" },
+    { word: "ゾ",   answer: "zo" },
+
+    { word: "ダ",   answer: "da" },
+    { word: "ヂ",   answer: "ji" },
+    { word: "ヅ",   answer: "zu" },
+    { word: "デ",   answer: "de" },
+    { word: "ド",   answer: "do" },
+
+    { word: "バ",   answer: "ba" },
+    { word: "ビ",   answer: "bi" },
+    { word: "ブ",   answer: "bu" },
+    { word: "ベ",   answer: "be" },
+    { word: "ボ",   answer: "bo" },
+
+    { word: "パ",   answer: "pa" },
+    { word: "ピ",   answer: "pi" },
+    { word: "プ",   answer: "pu" },
+    { word: "ペ",   answer: "pe" },
+    { word: "ポ",   answer: "po" },
+]