+++ src/vocabulary/domesticEquipment.js
@@ -1,24 +1,90 @@
 export default [
-    { word: "le lit", answer: "ベッド" },
-    { word: "le canapé", answer: ["ソファー", "ソファ"] },
-    { word: "le stylo", answer: "ペン" },
-    { word: "le stylo-bille", answer: "ボールペン" },
-    { word: "le cahier", answer: "ノート" },
-    { word: "le calendrier", answer: "カレンダー" },
-    { word: "la radio", answer: "ラジオ" },
-    { word: "la chaîne Hi-Fi", answer: ["ステレオ", "ステレオフォニック"] },
-    { word: "la télévision", answer: ["テレビ", "テレビジョン"] },
-    { word: "l'appareil photo", answer: "カメラ" },
-    { word: "l'ordinateur", answer: "コンピュータ" },
-    { word: "le PC", answer: ["パソコン", "パーソナルコンピュータ"] },
-    { word: "le climatiseur", answer: "エアコン" },
-    { word: "la télécommande", answer: "リモコン" },
-    { word: "la carte de paiement", answer: "クレジットカード" },
-    { word: "la porte", answer: "ドア" },
-    { word: "la lampe", answer: "ランプ" },
-    { word: "la table", answer: "ターブル" },
-    { word: "la fourchette", answer: "フォーク" },
-    { word: "le couteau", answer: "ナイフ" },
-    { word: "la cuillère", answer: "スプーン" },
-    { word: "les toilettes", answer: ["トイレ", "トイレット"] },
+    {
+        word: "le lit",
+        answer: "ベッド"
+    },
+    {
+        word: "le canapé",
+        answer: ["ソファー", "ソファ"]
+    },
+    {
+        word: "le stylo",
+        answer: "ペン"
+    },
+    {
+        word: "le stylo-bille",
+        answer: "ボールペン"
+    },
+    {
+        word: "le cahier",
+        answer: "ノート"
+    },
+    {
+        word: "le calendrier",
+        answer: "カレンダー"
+    },
+    {
+        word: "la radio",
+        answer: "ラジオ"
+    },
+    {
+        word: "la chaîne Hi-Fi",
+        answer: ["ステレオ", "ステレオフォニック"]
+    },
+    {
+        word: "la télévision",
+        answer: ["テレビ", "テレビジョン"]
+    },
+    {
+        word: "l'appareil photo",
+        answer: "カメラ"
+    },
+    {
+        word: "l'ordinateur",
+        answer: "コンピュータ"
+    },
+    {
+        word: "le PC",
+        answer: ["パソコン", "パーソナルコンピュータ"]
+    },
+    {
+        word: "le climatiseur",
+        answer: "エアコン"
+    },
+    {
+        word: "la télécommande",
+        answer: "リモコン"
+    },
+    {
+        word: "la carte de paiement",
+        answer: "クレジットカード"
+    },
+    {
+        word: "la porte",
+        answer: "ドア"
+    },
+    {
+        word: "la lampe",
+        answer: "ランプ"
+    },
+    {
+        word: "la table",
+        answer: "ターブル"
+    },
+    {
+        word: "la fourchette",
+        answer: "フォーク"
+    },
+    {
+        word: "le couteau",
+        answer: "ナイフ"
+    },
+    {
+        word: "la cuillère",
+        answer: "スプーン"
+    },
+    {
+        word: "les toilettes",
+        answer: ["トイレ", "トイレット"]
+    },
 ];