+++ src/vocabulary/index.js
@@ -1,3 +1,5 @@
+import countryNames from "./countryNames";
+import domesticAnimals from "./domesticAnimals";
 import domesticEquipment from "./domesticEquipment";
 import firstKanji from "./firstKanji";
 import numbers from "./numbers";
@@ -8,4 +10,6 @@ export default {
     "Premier Mots en Kanji": firstKanji,
     "Les Nombres": numbers,
     "Le Matériel Domestique": domesticEquipment,
+    "Le Nom des Pays": countryNames,
+    "Les Animaux Domestiques": domesticAnimals,
 };