+++ src/vocabulary/politeness.js
@@ -1,27 +1,102 @@
 export default [
-    { word: "oui", answer: "はい" },
-    { word: "non", answer: "いいえ" },
-    { word: "s'il vous plaît", answer: ["お願いします", "おねがいします"] },
-    { word: "merci", answer: ["ありがとう", "ありがとうございます"] },
-    { word: "il n'y a pas de quoi", answer: "どういたしまして" },
-    { word: "bonjour (le matin)", answer: ["おはよう", "おはようございます"] },
-    { word: "bonjour (la journée)", answer: "こんにちは" },
-    { word: "bonsoir", answer: "こんばんは" },
-    { word: "comment vas-tu (allez-vous)", answer: ["お元気ですか", "おゲンキですか"] },
-    { word: "oui, je vais bien (réponse)", answer: ["はい、元気です", "はい、ゲンキです"] },
-    { word: "au revoir", answer: ["さよなら", "またね"] },
-    { word: "bonne nuit", answer: ["おやすみ", "おやすみなさい"] },
-    { word: "à demain", answer: ["また明日", "またあした"] },
-    { word: "enchanté", answer: ["初めまして", "はじめまして", "どうぞ宜しくお願いします", "どうぞよろしくおねがいします"] },
-    { word: "je vous en prie", answer: "どうぞ" },
-    { word: "bon appétit", answer: "いただきます" },
-    { word: "excuse(z)-moi", answer: "すみません" },
-    { word: "pardonne(z)-moi", answer: ["ごめん", "ごめんなさい"] },
-    { word: "bon travail", answer: ["お疲れ様でした", "おつかれさまでした"] },
-    { word: "j'y vais", answer: "いってきます" },
-    { word: "bonne journée / bon départ", answer: "いってらっしゃい" },
-    { word: "je suis rentré", answer: "ただいま" },
-    { word: "bon retour", answer: ["おかえり", "おかえりなさい"] },
-    { word: "bienvenue", answer: ["いらっしゃいませ", "ようこそ"] },
-    { word: "je rentre (je dérange)", answer: ["お邪魔します", "おジャマします"] },
+    {
+        word: "oui",
+        answer: "はい"
+    },
+    {
+        word: "non",
+        answer: "いいえ"
+    },
+    {
+        word: "s'il vous plaît",
+        answer: ["お願いします", "おねがいします"]
+    },
+    {
+        word: "merci",
+        answer: ["ありがとう", "ありがとうございます"]
+    },
+    {
+        word: "il n'y a pas de quoi",
+        answer: "どういたしまして"
+    },
+    {
+        word: "bonjour (le matin)",
+        answer: ["おはよう", "おはようございます"]
+    },
+    {
+        word: "bonjour (la journée)",
+        answer: "こんにちは"
+    },
+    {
+        word: "bonsoir",
+        answer: "こんばんは"
+    },
+    {
+        word: "comment vas-tu (allez-vous)",
+        answer: ["お元気ですか", "おゲンキですか"]
+    },
+    {
+        word: "oui, je vais bien (réponse)",
+        answer: ["はい、元気です", "はい、ゲンキです"]
+    },
+    {
+        word: "au revoir",
+        answer: ["さようなら", "またね"]
+    },
+    {
+        word: "bonne nuit",
+        answer: ["おやすみ", "おやすみなさい"]
+    },
+    {
+        word: "à demain",
+        answer: ["また明日", "またあした"]
+    },
+    {
+        word: "enchanté",
+        answer: ["初めまして", "はじめまして", "どうぞ宜しくお願いします", "どうぞよろしくおねがいします"]
+    },
+    {
+        word: "je vous en prie",
+        answer: "どうぞ"
+    },
+    {
+        word: "bon appétit",
+        answer: "いただきます"
+    },
+    {
+        word: "excuse(z)-moi",
+        answer: "すみません"
+    },
+    {
+        word: "pardonne(z)-moi",
+        answer: ["ごめん", "ごめんなさい"]
+    },
+    {
+        word: "bon travail",
+        answer: ["お疲れ様でした", "おつかれさまでした"]
+    },
+    {
+        word: "j'y vais",
+        answer: "いってきます"
+    },
+    {
+        word: "bonne journée / bon départ",
+        answer: "いってらっしゃい"
+    },
+    {
+        word: "je suis rentré",
+        answer: "ただいま"
+    },
+    {
+        word: "bon retour",
+        answer: ["おかえり", "おかえりなさい"]
+    },
+    {
+        word: "bienvenue",
+        answer: ["いらっしゃいませ", "ようこそ"]
+    },
+    {
+        word: "je rentre (je dérange)",
+        answer: ["お邪魔します", "おジャマします"]
+    },
 ];